Use usize for GroupNum/GlobalWeek, eliminate unnecessary as casts
Reserve `as` strictly for ColloML Var interop (i32). Use try_into().expect()
at the i32→usize boundary, and i64::from() for ILP expressions.

+++ constraints-colloscopes/src/ids.rs
@@ -3,7 +3,7 @@ pub use collomatique_state_colloscopes::ids::{
 };
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct GlobalWeek(pub u32);
+pub struct GlobalWeek(pub usize);
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct GroupNum(pub u32);
+pub struct GroupNum(pub usize);
